Patch for game update 11 and exiled update 3

+++ RandomGamemode/Config.cs
@@ -30,12 +30,12 @@ namespace RandomGamemode
 		public int SCP682Health { get; private set; } = 5000;
 
 		[Description( "Amount of ammo MTF units should get for the SCP-682 Containment gamemode." )]
-		public uint SCP682MTFAmmo { get; private set; } = 1000;
+		public ushort SCP682MTFAmmo { get; private set; } = 1000;
 
 		[Description( "Max amount of dodgeballs that can be active in the world at once. Setting this too high will cause the server to hang." )]
 		public int MaxDodgeballs { get; private set; } = 20;
 
 		[Description( "Amount of ammo the scientist should get for the Night of the Living Nerd gamemode." )]
-		public uint NerdAmmoAmount { get; private set; } = 1000;
+		public ushort NerdAmmoAmount { get; private set; } = 1000;
 	}
 }
